# code/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kanban Exam Project",
        "contact": {
            "name": "kent vejrup madsen",
            "url": "https://github.com/KentVejrupMadsen",
            "email": "kent.vejrup.madsen@protonmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/KentVejrupMadsen/EASV.Exam.Kanban.Backend/blob/main/license.md"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend-exam-zu4kv.ondigitalocean.app",
            "description": "public available server"
        },
        {
            "url": "http://127.0.0.1",
            "description": "local development server"
        }
    ],
    "paths": {
...
        "/api/1.0.0/resources/images/create": {
            "post": {
                "tags": [
                    "1.0.0",
                    ""
                ],
                "operationId": "e9b935956e8fb24b59e55b72c021730f",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "bearer token - has to be included in the header of the request"
                    }
                ]
            }
        },
...
        "/api/1.0.0/resources/images/update": {
            "patch": {
                "tags": [
                    "1.0.0",
                    ""
                ],
                "operationId": "cf349aec98897e739de1236e8a250688",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "bearer token - has to be included in the header of the request"
                    }
                ]
            }
        },
        "/api/1.0.0/resources/images/delete/{id}": {
            "delete": {
                "tags": [
                    "1.0.0",
                    ""
                ],
                "operationId": "5bd66af2d308ea7a5d53526cbef63214",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "bearer token - has to be included in the header of the request"
                    }
                ]
            }
        },
...
        "/api/1.0.0/tasks/index": {
            "get": {
                "tags": [
                    "1.0.0",
                    ""
                ],
                "operationId": "cc7571e907bb446dc94f29b09c3f7fb5",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "bearer token - has to be included in the header of the request"
                    }
                ]
            }
        },
        "/api/1.0.0/tasks/create": {
            "post": {
                "tags": [
                    "1.0.0",
                    ""
                ],
                "operationId": "79ab373e4dc777413730ab69b3bd6433",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "bearer token - has to be included in the header of the request"
                    }
                ]
            }
        },
        "/api/1.0.0/tasks/groups/identity/{id}": {
            "get": {
                "tags": [
                    "1.0.0",
                    ""
                ],
                "operationId": "fe10f0c486d921f6d12ba21c29833e78",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "bearer token - has to be included in the header of the request"
                    }
                ]
            }
        },
        "/api/1.0.0/tasks/groups/update": {
            "post": {
                "tags": [
                    "1.0.0",
                    ""
                ],
                "operationId": "0b89577c1c8ccf916478acde2b8a8c75",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "has to be included in the header of the request"
                    }
                ]
            },
            "patch": {
                "tags": [
                    "1.0.0",
                    ""
                ],
                "operationId": "abe57dc4f7c7004b7d6961bc2146056c",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "bearer token - has to be included in the header of the request"
                    }
                ]
            }
        },
        "/api/1.0.0/tasks/groups/delete/{id}": {
            "delete": {
                "tags": [
                    "1.0.0",
                    ""
                ],
                "operationId": "2b2cfb62efae2cef52cf0ec865d2e197",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "bearer token - has to be included in the header of the request"
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "bearer token - has to be included in the header of the request",
                "name": "bearer authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}